backend/pyproject.toml
[project]
name = "document-ai-workbench-api"
version = "0.1.0"
description = "FastAPI backend for the multimodal document AI workbench."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "pydantic-settings",
  "pymupdf",
  "python-multipart",
  "qdrant-client",
  "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
  "httpx",
  "pytest",
  "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
